refactored tile drawer and tile-selector to remove duplicates
