+++ webapp/src/app/services/phaser/tilemap/tile-drawer.ts
@@ -9,22 +9,16 @@ import { CCMapLayer } from './cc-map-layer';
 import { Filler } from './fill';
 import { pointsInLine } from './points-in-line';
 import { customPutTileAt } from './layer-helper';
+import { BaseTileDrawer } from '../BaseTileDrawer';
 
 export class TileDrawer extends BaseObject {
 	
-	private layer?: CCMapLayer;
 	private selectedTiles: SelectedTile[] = [];
+	private layer?: CCMapLayer;
 	
-	private selection?: Phaser.GameObjects.Container;
+	private baseDrawer!: BaseTileDrawer;
 	
-	private previewTileMap!: Phaser.Tilemaps.Tilemap;
-	private previewLayer?: Phaser.Tilemaps.TilemapLayer;
-	
-	private rightClickStart?: Point;
-	private rightClickEnd?: Point;
 	private renderLayersTransparent = false;
-	private rightPointerDown = false;
-	
 	
 	private container!: Phaser.GameObjects.Container;
 	
@@ -40,7 +34,9 @@ export class TileDrawer extends BaseObject {
 		super(scene, 'tileDrawer');
 	}
 	
-	protected init() {
+	
+	protected override init() {
+		
 		this.fillKey = this.scene.input.keyboard!.addKey(Phaser.Input.Keyboard.KeyCodes.F, false);
 		this.transparentKey = this.scene.input.keyboard!.addKey(Phaser.Input.Keyboard.KeyCodes.R, false);
 		this.visibilityKey = this.scene.input.keyboard!.addKey(Phaser.Input.Keyboard.KeyCodes.V, false);
@@ -49,15 +45,14 @@ export class TileDrawer extends BaseObject {
 		this.container = this.scene.add.container(0, 0);
 		this.container.depth = 10000;
 		
-		this.drawRect(1, 1);
-		
-		this.resetSelectedTiles();
-		
-		this.previewTileMap = this.scene.add.tilemap(undefined, Globals.TILE_SIZE, Globals.TILE_SIZE);
+		this.baseDrawer = new BaseTileDrawer(this.scene, false, this.container);
+		this.baseDrawer.resetSelectedTiles();
+		this.scene.add.existing(this.baseDrawer);
 	}
 	
-	private selectLayer(selectedLayer?: CCMapLayer) {
+	private async selectLayer(selectedLayer?: CCMapLayer) {
 		this.layer = selectedLayer;
+		await this.baseDrawer.setLayer(selectedLayer);
 		
 		this.setLayerAlpha();
 		
@@ -66,31 +61,7 @@ export class TileDrawer extends BaseObject {
 			return;
 		}
 		this.container.visible = true;
-		const tileset = this.previewTileMap.addTilesetImage('only', selectedLayer.details.tilesetName);
-		if (tileset) {
-			tileset.firstgid = 1;
-		}
 	}
-	
-	private updateSelectedTiles(selected: SelectedTile[]) {
-		this.selectedTiles = selected;
-		this.renderPreview();
-		
-		let x = 0;
-		let y = 0;
-		selected.forEach(tile => {
-			const o = tile.offset;
-			if (o.x > x) {
-				x = o.x;
-			}
-			if (o.y > y) {
-				y = o.y;
-			}
-		});
-		
-		this.drawRect(x + 1, y + 1, 0, 0);
-	}
-	
 	
 	preUpdate(): void {
 		// hide cursor when no map loaded
@@ -100,54 +71,6 @@ export class TileDrawer extends BaseObject {
 		}
 		const pointer = this.scene.input.activePointer;
 		const p = Helper.worldToTile(pointer.worldX - this.layer.x, pointer.worldY - this.layer.y);
-		
-		// render selection border
-		if (this.rightClickStart) {
-			Helper.clampToBounds(this.layer, p);
-			
-			if (this.rightClickEnd && this.rightClickEnd.x === p.x && this.rightClickEnd.y === p.y) {
-				// shortcut to avoid redrawing rectangle every frame
-				return;
-			}
-			
-			this.rightClickEnd = p;
-			const diff = Vec2.sub(p, this.rightClickStart, true);
-			const start = {x: 0, y: 0};
-			if (diff.x >= 0) {
-				diff.x++;
-			} else {
-				start.x = Globals.TILE_SIZE;
-				diff.x--;
-			}
-			if (diff.y >= 0) {
-				diff.y++;
-			} else {
-				start.y = Globals.TILE_SIZE;
-				diff.y--;
-			}
-			
-			this.drawRect(diff.x, diff.y, start.x, start.y, true);
-			return;
-		}
-		
-		// position tile drawer border to cursor
-		const container = this.container;
-		container.x = pointer.worldX;
-		container.y = pointer.worldY;
-		
-		if (container.x < this.layer.x) {
-			container.x -= Globals.TILE_SIZE;
-		}
-		if (container.y < this.layer.y) {
-			container.y -= Globals.TILE_SIZE;
-		}
-		
-		container.x -= (container.x - this.layer.x) % Globals.TILE_SIZE;
-		container.y -= (container.y - this.layer.y) % Globals.TILE_SIZE;
-		
-		if (this.previewLayer) {
-			Vec2.assign(this.previewLayer, container);
-		}
 		
 		// draw tiles
 		// trigger only when mouse is over canvas element (the renderer), avoids triggering when interacting with ui
@@ -197,36 +120,13 @@ export class TileDrawer extends BaseObject {
 	
 	protected deactivate() {
 		this.container.visible = false;
-		if (this.previewLayer) {
-			this.previewLayer.visible = false;
-		}
+		this.baseDrawer.setActive(false);
 	}
 	
 	protected activate() {
-		if (this.previewLayer) {
-			this.previewLayer.visible = true;
-		}
-		const sub = Globals.mapLoaderService.selectedLayer.subscribe(layer => this.selectLayer(layer));
-		this.addSubscription(sub);
-		
-		const sub2 = Globals.phaserEventsService.changeSelectedTiles.subscribe(tiles => this.updateSelectedTiles(tiles));
-		this.addSubscription(sub2);
-		
-		const pointerDown = (pointer: Phaser.Input.Pointer) => {
-			if (pointer.rightButtonDown()) {
-				this.onMouseRightDown();
-			}
-		};
-		this.addKeybinding({event: 'pointerdown', fun: pointerDown, emitter: this.scene.input});
-		
-		const pointerUp = (pointer: Phaser.Input.Pointer) => {
-			if (pointer.rightButtonReleased() && this.rightPointerDown) {
-				this.onMouseRightUp();
-			}
-		};
-		this.addKeybinding({event: 'pointerup', fun: pointerUp, emitter: this.scene.input});
-		this.addKeybinding({event: 'pointerupoutside', fun: pointerUp, emitter: this.scene.input});
-		
+		this.addSubscription(Globals.mapLoaderService.selectedLayer.subscribe(layer => this.selectLayer(layer)));
+		this.addSubscription(Globals.phaserEventsService.changeSelectedTiles.subscribe(tiles => this.selectedTiles = tiles));
+		this.baseDrawer.setActive(true);
 		
 		const fill = () => {
 			if (!Helper.isInputFocused()) {
@@ -286,154 +186,6 @@ export class TileDrawer extends BaseObject {
 				this.layer.alpha = 1;
 			}
 		}
-	}
-	
-	private onMouseRightDown() {
-		this.rightPointerDown = true;
-		if (!this.layer) {
-			return;
-		}
-		
-		
-		// only start tile copy when cursor in bounds
-		const pointer = this.scene.input.activePointer;
-		const p = Helper.worldToTile(pointer.worldX - this.layer.x, pointer.worldY - this.layer.y);
-		if (!Helper.isInBounds(this.layer, p)) {
-			return;
-		}
-		
-		this.resetSelectedTiles();
-		this.renderPreview();
-		this.rightClickStart = p;
-	}
-	
-	private drawRect(width: number, height: number, x = 0, y = 0, renderSize = false) {
-		if (this.selection) {
-			this.selection.destroy();
-		}
-		
-		let textColor = 'rgba(0,0,0,0.6)';
-		let backgroundColor = 0xffffff;
-		if (Globals.settingsService.getSettings().selectionBoxDark) {
-			textColor = 'rgba(255,255,255,0.9)';
-			backgroundColor = 0x333333;
-		}
-		
-		this.selection = this.scene.add.container(x, y);
-		
-		const rect = this.scene.add.rectangle(0, 0, width * Globals.TILE_SIZE, height * Globals.TILE_SIZE);
-		rect.setOrigin(0, 0);
-		rect.setStrokeStyle(1, backgroundColor, 0.6);
-		
-		this.selection.add(rect);
-		this.container.add(this.selection);
-		
-		if (!renderSize) {
-			Globals.globalEventsService.updateTileSelectionSize.next(undefined);
-			return;
-		}
-		
-		const makeText = (pos: Point, val: number) => {
-			const text = this.scene.add.text(pos.x, pos.y, Math.abs(val) + '', {
-				font: '400 10px Roboto',
-				color: textColor,
-				resolution: window.devicePixelRatio * 3,
-			});
-			text.setOrigin(0.5, 0);
-			const background = this.scene.add.rectangle(pos.x, pos.y + 2, 14, 10, backgroundColor, 0.6);
-			background.setOrigin(0.5, 0);
-			
-			this.selection?.add(background);
-			this.selection?.add(text);
-		};
-		
-		if (Math.abs(width) >= 3) {
-			makeText({
-				x: width * Globals.TILE_SIZE / 2,
-				y: (height > 0 ? 0 : height * Globals.TILE_SIZE) - 1
-			}, width);
-		}
-		
-		if (Math.abs(height) >= 3) {
-			makeText({
-				x: Globals.TILE_SIZE / 2 + (width > 0 ? 0 : width * Globals.TILE_SIZE),
-				y: (height - 1) * Globals.TILE_SIZE / 2,
-			}, height);
-		}
-		
-		Globals.globalEventsService.updateTileSelectionSize.next({
-			x: Math.abs(width),
-			y: Math.abs(height)
-		});
-	}
-	
-	private onMouseRightUp() {
-		this.rightPointerDown = false;
-		if (!this.layer) {
-			return;
-		}
-		this.selectedTiles = [];
-		
-		// cancel current selection when out of bounds
-		const phaserLayer = this.layer.getPhaserLayer();
-		if (!this.rightClickStart || !this.rightClickEnd || !phaserLayer) {
-			this.drawRect(1, 1);
-			this.resetSelectedTiles();
-			this.renderPreview();
-			return;
-		}
-		
-		// select tiles
-		const start = this.rightClickStart;
-		const end = this.rightClickEnd;
-		
-		const smaller = {
-			x: Math.min(start.x, end.x),
-			y: Math.min(start.y, end.y)
-		};
-		
-		const bigger = {
-			x: Math.max(start.x, end.x),
-			y: Math.max(start.y, end.y)
-		};
-		
-		const width = bigger.x - smaller.x + 1;
-		const height = bigger.y - smaller.y + 1;
-		
-		const tilesWithin = phaserLayer.getTilesWithin(smaller.x, smaller.y, width, height);
-		
-		const tiles: SelectedTile[] = tilesWithin.map(tile => ({
-			id: tile.index,
-			offset: Vec2.sub(tile, smaller, true)
-		}));
-		
-		Globals.phaserEventsService.changeSelectedTiles.next(tiles);
-		
-		this.rightClickStart = undefined;
-		this.rightClickEnd = undefined;
-	}
-	
-	private renderPreview() {
-		
-		// reset last draw when selected tiles change
-		this.lastDraw.x = -1;
-		this.previewTileMap.removeAllLayers();
-		const layer = this.previewTileMap.createBlankLayer('layer', 'only', 0, 0, 40, 40)!;
-		
-		this.selectedTiles.forEach(tile => {
-			customPutTileAt(tile.id, tile.offset.x, tile.offset.y, layer.layer);
-		});
-		
-		this.previewLayer = layer;
-		this.previewLayer.depth = this.container.depth - 1;
-		this.previewLayer.alpha = 0.6;
-		
-		// TODO: phaser bug fix, see https://github.com/photonstorm/phaser/issues/4642
-		this.previewTileMap.layers = [this.previewLayer.layer];
-	}
-	
-	private resetSelectedTiles() {
-		this.selectedTiles = [{id: 0, offset: {x: 0, y: 0}}];
 	}
 	
 	private fill() {
